Update for compatibility with systemd 233 and later
systemd 233 introduced the `system-update-cleanup.service` unit, which
is meant to recover from a failed system update. However,
`system-update-pacman.service`'s previous configuration was not
compatible with this unit: the cleanup service would run concurrently
with the system update, rebooting the system before the update was
complete. The new configuration introduced in this unit solves the
issue.